+++ src/Blob/Models/ListPageBlobRangesOptions.php
@@ -25,6 +25,7 @@
 namespace MicrosoftAzure\Storage\Blob\Models;
 
 use MicrosoftAzure\Storage\Common\Internal\Validate;
+use MicrosoftAzure\Storage\Common\Models\Range;
 
 /**
  * Optional parameters for listPageBlobRanges wrapper
@@ -38,7 +39,8 @@ use MicrosoftAzure\Storage\Common\Internal\Validate;
  */
 class ListPageBlobRangesOptions extends BlobServiceOptions
 {
-    private $_snapshot;
+    private $snapshot;
+    private $range;
     private $_rangeStart;
     private $_rangeEnd;
     
@@ -49,7 +51,7 @@ class ListPageBlobRangesOptions extends BlobServiceOptions
      */
     public function getSnapshot()
     {
-        return $this->_snapshot;
+        return $this->snapshot;
     }
 
     /**
@@ -61,52 +63,28 @@ class ListPageBlobRangesOptions extends BlobServiceOptions
      */
     public function setSnapshot($snapshot)
     {
-        $this->_snapshot = $snapshot;
+        $this->snapshot = $snapshot;
     }
-    
+
     /**
-     * Gets rangeStart
+     * Gets Blob range.
      *
-     * @return integer
+     * @return Range
      */
-    public function getRangeStart()
+    public function getRange()
     {
-        return $this->_rangeStart;
+        return $this->range;
     }
-    
+
     /**
-     * Sets rangeStart
+     * Sets Blob range.
      *
-     * @param integer $rangeStart the blob lease id.
+     * @param Range $range value.
      *
      * @return void
      */
-    public function setRangeStart($rangeStart)
+    public function setRange(Range $range)
     {
-        Validate::isInteger($rangeStart, 'rangeStart');
-        $this->_rangeStart = $rangeStart;
-    }
-    
-    /**
-     * Gets rangeEnd
-     *
-     * @return integer
-     */
-    public function getRangeEnd()
-    {
-        return $this->_rangeEnd;
-    }
-    
-    /**
-     * Sets rangeEnd
-     *
-     * @param integer $rangeEnd range end value in bytes
-     *
-     * @return void
-     */
-    public function setRangeEnd($rangeEnd)
-    {
-        Validate::isInteger($rangeEnd, 'rangeEnd');
-        $this->_rangeEnd = $rangeEnd;
+        $this->range = $range;
     }
 }